requirements_file.txt
streamlit==1.28.1
streamlit-autorefresh==0.0.1
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
plotly==5.15.0
pytz==2023.3
openpyxl==3.1.2
python-dotenv==1.0.0